Fix for flick not working on uiautomator2
[Problem]
Flick command on any element does not work with uiautomator2.

[Cause]
Flick command is proxied to Uiautomator2 server running on device.
The server expects command body to have element identifier with keyname "elementId".
But the command is proxied with body having element identifier with key "element".
This was causing uiautomator to not identify element and hence the failure of command.

[Solution]
Added code to replace text "element" with "elementId:" for proxyReqRes commands
before sending command to the proxy server.

[Test]
Verified that flick works after the fix